SwiftUI

+++ SwiftUIBootcamp/MemoTwice/MemoTwice/Views/ListView.swift
@@ -12,21 +12,27 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear(duration: 0.05)) {
-                            listViewModel.updateItem(item: item)
-                            print(item.id)
-                        }
+        ZStack {
+            if listViewModel.items.isEmpty {
+                NoItemsView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+            } else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear(duration: 0.05)) {
+                                    listViewModel.updateItem(item: item)
+                                    print(item.id)
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
+                .listStyle(.plain)
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
-            
         }
-        .listStyle(.plain)
         .navigationTitle("Todo List📝")
         .navigationBarItems(leading: EditButton(),
                             trailing: NavigationLink("Add", destination: AddView() ))
